Work with filename state in only doc tree.
